Fixed handsani_tower
Handsani_tower now deals 10 damage to 5 enemies before disappearing!!! This took me so long to figure out but I flippin' did it $$$$

In an ideal world, I would also want the sprite to become increasingly transparent as it gets used up or "evaporates" heeehee. But it's not that important - it would be cool though!